test(proxy): add streaming+cache integration coverage (#598)
integration_slot_roundtrip.rs only exercised the non-streaming
cache_lifecycle::run_with_cache cycle; the streaming path through
sse_stream::spawn_and_return had no coverage of save-after-generation,
restore-before-generation, or SSE frame integrity under cache hooks.

Consolidates the FixedUpstream/TaggedCatalog mocks and mock-upstream/
SSE-assertion helpers that integration_slot_roundtrip.rs and
integration_proxy_pipeline.rs each hand-rolled into tests/fixtures/common.rs,
and adds the new streaming+cache test on top of the shared helpers.
